# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Workplace English",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/sessions": {
      "get": {
        "summary": "Sessions",
        "operationId": "sessions_api_sessions_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/Session"
                  },
                  "type": "array",
                  "title": "Response Sessions Api Sessions Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create",
        "operationId": "create_api_sessions_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSession"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Session"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/sessions/{identifier}": {
      "get": {
        "summary": "Get Session",
        "operationId": "get_session_api_sessions__identifier__get",
        "parameters": [
          {
            "name": "identifier",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[a-f0-9]{32}$",
              "title": "Identifier"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Session"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete",
        "operationId": "delete_api_sessions__identifier__delete",
        "parameters": [
          {
            "name": "identifier",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[a-f0-9]{32}$",
              "title": "Identifier"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/sessions/{identifier}/connect": {
      "post": {
        "summary": "Connect",
        "operationId": "connect_api_sessions__identifier__connect_post",
        "parameters": [
          {
            "name": "identifier",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[a-f0-9]{32}$",
              "title": "Identifier"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConnectResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/sessions/{identifier}/commands": {
      "post": {
        "summary": "Apply Command",
        "operationId": "apply_command_api_sessions__identifier__commands_post",
        "parameters": [
          {
            "name": "identifier",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[a-f0-9]{32}$",
              "title": "Identifier"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Command-Input"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Session"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/sessions/{identifier}/heartbeat": {
      "post": {
        "summary": "Heartbeat",
        "operationId": "heartbeat_api_sessions__identifier__heartbeat_post",
        "parameters": [
          {
            "name": "identifier",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[a-f0-9]{32}$",
              "title": "Identifier"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Heartbeat"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/sessions/{identifier}/events": {
      "get": {
        "summary": "Events",
        "operationId": "events_api_sessions__identifier__events_get",
        "parameters": [
          {
            "name": "identifier",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[a-f0-9]{32}$",
              "title": "Identifier"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/sessions/{identifier}/playback": {
      "post": {
        "summary": "Playback",
        "operationId": "playback_api_sessions__identifier__playback_post",
        "parameters": [
          {
            "name": "identifier",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[a-f0-9]{32}$",
              "title": "Identifier"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Playback"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/content/{purpose_id}/playback": {
      "post": {
        "summary": "Expression Playback",
        "operationId": "expression_playback_api_content__purpose_id__playback_post",
        "parameters": [
          {
            "name": "purpose_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Purpose Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExpressionPlayback"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "CoachingTurn": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "input_id": {
            "type": "string",
            "title": "Input Id"
          },
          "transcript": {
            "$ref": "#/components/schemas/Transcript"
          },
          "reply": {
            "$ref": "#/components/schemas/CoachingReply"
          },
          "message_id": {
            "type": "string",
            "title": "Message Id"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "input_id",
          "transcript",
          "reply",
          "message_id"
        ],
        "title": "CoachingTurn"
      },
      "Command-Input": {
        "properties": {
          "command_id": {
            "type": "string",
            "pattern": "^[a-f0-9]{32}$",
            "title": "Command Id"
          },
          "expected_revision": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Expected Revision"
          },
          "connection_epoch": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Connection Epoch"
          },
          "type": {
            "type": "string",
            "enum": [
              "pause",
              "finish",
              "done",
              "use_answer",
              "try_again",
              "retry_operation",
              "mute",
              "unmute",
              "replay",
              "continue_answer",
              "hint",
              "focused_retry",
              "expression"
            ],
            "title": "Type"
          },
          "input_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Input Id"
          },
          "expression_id": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "expression",
                  "alternative",
                  "example"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Expression Id"
          },
          "target_note_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target Note Id"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "command_id",
          "expected_revision",
          "type"
        ],
        "title": "Command"
      },
...
      "ConnectResult": {
        "properties": {
          "session": {
            "$ref": "#/components/schemas/Session"
          },
          "url": {
            "type": "string",
            "title": "Url"
          },
          "token": {
            "type": "string",
            "title": "Token"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "session",
          "url",
          "token"
        ],
        "title": "ConnectResult"
      },
...
      "CreateSession": {
        "properties": {
          "purpose_id": {
            "type": "string",
            "title": "Purpose Id"
          },
          "command_id": {
            "type": "string",
            "pattern": "^[a-f0-9]{32}$",
            "title": "Command Id"
          },
          "source_session_id": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^[a-f0-9]{32}$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Session Id"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "purpose_id",
          "command_id"
        ],
        "title": "CreateSession"
      },
...
      "ExpressionPlayback": {
        "properties": {
          "expression_id": {
            "type": "string",
            "enum": [
              "expression",
              "alternative",
              "example"
            ],
            "title": "Expression Id"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "expression_id"
        ],
        "title": "ExpressionPlayback"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "Heartbeat": {
        "properties": {
          "connection_epoch": {
            "type": "string",
            "title": "Connection Epoch"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "connection_epoch"
        ],
        "title": "Heartbeat"
      },
...
      "Playback": {
        "properties": {
          "message_id": {
            "type": "string",
            "title": "Message Id"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "message_id"
        ],
        "title": "Playback"
      },
...
      "Session": {
        "properties": {
          "schema_version": {
            "type": "integer",
            "title": "Schema Version",
            "default": 1
          },
          "id": {
            "type": "string",
            "title": "Id"
          },
          "guest_id": {
            "type": "string",
            "title": "Guest Id"
          },
          "purpose_id": {
            "type": "string",
            "title": "Purpose Id"
          },
          "content_version": {
            "type": "string",
            "title": "Content Version"
          },
          "conversation_model": {
            "type": "string",
            "title": "Conversation Model"
          },
          "prompt_version": {
            "type": "string",
            "title": "Prompt Version",
            "default": "conversation-v1"
          },
          "revision": {
            "type": "integer",
            "title": "Revision",
            "default": 1
          },
          "sequence": {
            "type": "integer",
            "title": "Sequence",
            "default": 1
          },
          "lifecycle": {
            "type": "string",
            "enum": [
              "in_progress",
              "paused",
              "completed"
            ],
            "title": "Lifecycle",
            "default": "paused"
          },
          "phase": {
            "type": "string",
            "enum": [
              "roleplay",
              "coaching",
              "retry"
            ],
            "title": "Phase",
            "default": "roleplay"
          },
          "substate": {
            "type": "string",
            "enum": [
              "connecting",
              "opening",
              "listening",
              "thinking",
              "speaking",
              "capped",
              "assessment_pending",
              "ready",
              "unavailable"
            ],
            "title": "Substate",
            "default": "connecting"
          },
          "created_at": {
            "type": "number",
            "title": "Created At"
          },
          "last_practice_at": {
            "type": "number",
            "title": "Last Practice At"
          },
          "expires_at": {
            "type": "number",
            "title": "Expires At"
          },
          "completed_at": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "connection_epoch": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Connection Epoch"
          },
          "room": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Room"
          },
          "participant": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Participant"
          },
          "generation_id": {
            "type": "string",
            "title": "Generation Id"
          },
          "situation": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Situation"
          },
          "setup": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Setup"
          },
          "question_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Question Id"
          },
          "answers": {
            "items": {
              "$ref": "#/components/schemas/Answer"
            },
            "type": "array",
            "maxItems": 3,
            "title": "Answers"
          },
          "messages": {
            "items": {
              "$ref": "#/components/schemas/TutorMessage"
            },
            "type": "array",
            "maxItems": 100,
            "title": "Messages"
          },
          "input": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Input"
              },
              {
                "type": "null"
              }
            ]
          },
          "pending_command": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Command-Output"
              },
              {
                "type": "null"
              }
            ]
          },
          "receipts": {
            "additionalProperties": {
              "$ref": "#/components/schemas/Receipt"
            },
            "type": "object",
            "title": "Receipts"
          },
          "assessment": {
            "$ref": "#/components/schemas/Assessment"
          },
          "notice": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notice"
          },
          "capture_muted": {
            "type": "boolean",
            "title": "Capture Muted",
            "default": false
          },
          "replay_restore_capture": {
            "type": "boolean",
            "title": "Replay Restore Capture",
            "default": false
          },
          "helper": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Helper"
              },
              {
                "type": "null"
              }
            ]
          },
          "retries": {
            "items": {
              "$ref": "#/components/schemas/FocusedRetry"
            },
            "type": "array",
            "maxItems": 12,
            "title": "Retries"
          },
          "coaching_turns": {
            "items": {
              "$ref": "#/components/schemas/CoachingTurn"
            },
            "type": "array",
            "maxItems": 30,
            "title": "Coaching Turns"
          },
          "source_session_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Session Id"
          },
          "previous_opening": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Previous Opening"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "schema_version",
          "id",
          "guest_id",
          "purpose_id",
          "content_version",
          "conversation_model",
          "prompt_version",
          "revision",
          "sequence",
          "lifecycle",
          "phase",
          "substate",
          "created_at",
          "last_practice_at",
          "expires_at",
          "completed_at",
          "connection_epoch",
          "room",
          "participant",
          "generation_id",
          "situation",
          "setup",
          "question_id",
          "answers",
          "messages",
          "input",
          "pending_command",
          "receipts",
          "assessment",
          "notice",
          "capture_muted",
          "replay_restore_capture",
          "helper",
          "retries",
          "coaching_turns",
          "source_session_id",
          "previous_opening"
        ],
        "title": "Session"
      },
...
    }
  }
}